Supp 2
This is the script used for the generation of the regression model. It produces a plot that displays the regressor's prediction performance compared to the actual Rosetta scores. this script can also be used to create a classifier model. Both models return performance indicators such as area under the curve, r2 values, absolute mean error, classification reports, and their confusion matrix associate with the data set.