Supp 15
This is a composition of all the consensus sequences for each cluster, for every encoder used. these are compared to one another and the total consensus to more easily illustrate differences between the encoders. One Hot encode was chosen to have a 4 cluster arrangement based off its dendrogram which is why it has an additional cluster in comparison to the other two encoders