Supp 12
This is the scatter plot based off the initial clustering assignment from the Blosum62 encoded sequences, represented by the first and second principle components for the 2D plot, and the first, second, and third clusters for the 3D plot. This was generated during the preliminary round of data analysis and therefore only contains the three clusters for One-Hot.